🎨 Better message forwarding logic
Instead of refraining from forwarding messages based on the potential recipient's
IP address, which could be the same across several devices or apps on a single device,
calculate a hash of each message and ensure that each message is forwarded only once.
Legitimate unique messages from peers will be signed with unique signatures, so peers
should still be able to send messages to each other with the same contents.